requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Mail==0.9.1
python-dotenv==1.0.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
